vite

+++ database/migrations/2025_03_06_134750_create_leaderboard_table.php
@@ -11,18 +11,20 @@ return new class extends Migration
      */
     public function up()
 {
-    Schema::create('leaderboard', function (Blueprint $table) {
-        $table->id();
-        $table->foreignId('user_id')->nullable()->constrained('users')->onDelete('cascade'); // ✅ السماح بـ NULL
-        $table->string('name');
-        $table->integer('score')->default(0);
-        $table->integer('wins')->default(0);
-        $table->integer('losses')->default(0);
-        $table->timestamp('last_match_time')->nullable();
-        $table->integer('crystals')->default(0);
-        $table->timestamps();
-    });
-    
+    if (!Schema::hasTable('leaderboard')) {
+        Schema::create('leaderboard', function (Blueprint $table) {
+            $table->id();
+            $table->foreignId('user_id')->nullable()->constrained('users')->onDelete('cascade');
+            $table->string('name');
+            $table->integer('score')->default(0);
+            $table->integer('wins')->default(0);
+            $table->integer('losses')->default(0);
+            $table->timestamp('last_match_time')->nullable();
+            $table->integer('crystals')->default(0);
+            $table->timestamps();
+        });
+    }
 }
+
 
 };
